refactor: or-3126 move PowerBI exporters to seperate file
